+++ db/migrate/20151018202245_add_weighted_score_to_comments.rb
@@ -1,5 +1,5 @@
 class AddWeightedScoreToComments < ActiveRecord::Migration
   def change
-    add_column :comments, :weighted_score, :integer, default: 0
+    add_column :comments, :weighted_score, :float, default: 0
   end
 end
